Rename `NavigatorRoot` to  `NavigatorProvider`
