+++ docs/manifest.json
@@ -1068,9 +1068,9 @@
 		"parent": "components"
 	},
 	{
-		"title": "NavigatorRoot",
-		"slug": "navigator-root",
-		"markdown_source": "../packages/components/src/navigator/navigator-root/README.md",
+		"title": "NavigatorProvider",
+		"slug": "navigator-provider",
+		"markdown_source": "../packages/components/src/navigator/navigator-provider/README.md",
 		"parent": "components"
 	},
 	{
