+++ packages/components/src/navigator/index.js
@@ -1,3 +1,3 @@
-export { default as NavigatorRoot } from './navigator-root';
+export { default as NavigatorProvider } from './navigator-provider';
 export { default as NavigatorScreen } from './navigator-screen';
 export { default as useNavigator } from './use-navigator';
