+++ packages/components/src/navigator/navigator-provider/index.js
